requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
click==8.3.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.124.4
h11==0.16.0
idna==3.11
mongoengine==0.29.1
passlib==1.7.4
pip==24.0
pydantic==2.12.5
pydantic-core==2.41.5
pymongo==4.15.5
python-dotenv==1.2.1
redis==7.1.0
starlette==0.50.0
typing-extensions==4.15.0
typing-inspection==0.4.2
uv==0.9.17
uvicorn==0.38.0
